fix: extract useDrawer hook to deduplicate mobile nav drawers
Shared focus-trap, body-lock, inert toggle, and keyboard logic
extracted into hooks/use-drawer.ts. Fixes Sonar duplicated-lines
warning between dashboard and landing mobile-nav-drawer components.
Also fixes sidebar named-export warning (version import).
