+++ frontend/src/components/landing/mobile-nav-drawer.tsx
@@ -2,9 +2,9 @@
 
 import Link from "next/link";
 import { Menu, Xmark } from "iconoir-react";
-import { useCallback, useEffect, useRef, useState } from "react";
 
 import { buttonVariants } from "@/components/ui/button";
+import { useDrawer } from "@/hooks/use-drawer";
 import { cn } from "@/lib/cn";
 
 const LINKS: ReadonlyArray<{ label: string; href: string }> = [
@@ -13,79 +13,7 @@ const LINKS: ReadonlyArray<{ label: string; href: string }> = [
 ];
 
 export function MobileNavDrawer({ scrolled }: Readonly<{ scrolled: boolean }>) {
-  const [open, setOpen] = useState(false);
-  const drawerRef = useRef<HTMLDialogElement | null>(null);
-  const triggerRef = useRef<HTMLButtonElement | null>(null);
-  const mainRef = useRef<HTMLElement | null>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    const prev = document.body.style.overflow;
-    document.body.style.overflow = "hidden";
-    return () => {
-      document.body.style.overflow = prev;
-    };
-  }, [open]);
-
-  useEffect(() => {
-    if (!open) return;
-    const main = document.querySelector("main");
-    if (main) {
-      main.setAttribute("inert", "");
-      mainRef.current = main;
-    }
-    return () => {
-      if (mainRef.current) {
-        mainRef.current.removeAttribute("inert");
-        mainRef.current = null;
-      }
-    };
-  }, [open]);
-
-  useEffect(() => {
-    if (!open) return;
-    const onKey = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        setOpen(false);
-        triggerRef.current?.focus();
-        return;
-      }
-      if (event.key === "Tab") {
-        const drawer = drawerRef.current;
-        if (!drawer) return;
-        const focusable = drawer.querySelectorAll<HTMLElement>(
-          'a[href], button:not([disabled])',
-        );
-        if (focusable.length === 0) return;
-        const first = focusable[0]!;
-        const last = focusable[focusable.length - 1]!;
-        if (event.shiftKey) {
-          if (document.activeElement === first) {
-            event.preventDefault();
-            last.focus();
-          }
-        } else if (document.activeElement === last) {
-          event.preventDefault();
-          first.focus();
-        }
-      }
-    };
-    globalThis.addEventListener("keydown", onKey);
-    return () => globalThis.removeEventListener("keydown", onKey);
-  }, [open]);
-
-  useEffect(() => {
-    if (!open) return;
-    const first = drawerRef.current?.querySelector<HTMLElement>(
-      'a[href], button:not([disabled])',
-    );
-    first?.focus();
-  }, [open]);
-
-  const close = useCallback(() => {
-    setOpen(false);
-    triggerRef.current?.focus();
-  }, []);
+  const { open, setOpen, close, drawerRef, triggerRef } = useDrawer();
 
   return (
     <>
